+++ packages/hooks_riverpod/lib/misc.dart
@@ -20,4 +20,5 @@ export 'src/internals.dart'
         ProviderListenable,
         Override,
         ProviderBase,
-        KeepAliveLink;
+        KeepAliveLink,
+        ProviderException;