+++ packages/riverpod/test/old/legacy_providers/stream_provider/stream_provider_test.dart
@@ -352,7 +352,7 @@ void main() {
       await expectLater(
         future,
         throwsA(
-          isA<StateError>().having(
+          isStateError.having(
             (e) => e.message,
             'message',
             equalsIgnoringHashCodes(